pyproject.toml
[project]
name = "finworld"
version = "0.1.0"
description = "FinWorld"
authors = [
    {name = "finworld",email = "example@gmail.com"}
]
license = {text = "MIT LICENSE"}
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry.dependencies]
mmengine = "^0.10.7"
python-dotenv = "^1.0.1"
transformers = "^4.49.0"
joblib = "^1.4.2"
pandas = "^2.2.3"
trl = "^0.15.2"
wandb = "^0.19.11"
numpy = "^1.26.4"
tensorboard = "^2.19.0"
timm = "^1.0.15"
diffusers = "^0.32.2"
scipy = "^1.15.2"
vit-pytorch = "^1.10.1"
ftfy = "^6.3.1"
tiktoken = "^0.9.0"
gym = "^0.26.2"
ninja = "^1.11.1.3"
modelscope = "^1.24.0"
ray = "^2.44.0"
pypdf2 = "^3.0.1"
playwright = "^1.51.0"
colorama = "^0.4.6"
openai = "^1.61.0"
duckduckgo-search = "^7.5.2"
litellm = "1.64.1"
pandas-market-calendars = "^4.6.1"
markdownify = "^1.1.0"
fake-useragent = "^2.1.0"
selenium = "^4.30.0"
webdriver-manager = "^4.0.2"
playwright-stealth = "^1.0.6"
codetiming = "^1.4.0"
omegaconf = "^2.3.0"
torchdata = "^0.11.0"
hydra-core = "^1.3.2"
tabulate = "^0.9.0"
mathruler = "^0.1.0"
pylatexenc = "^2.10"
alpaca-py = "^0.40.1"
camelot-py = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
